[backend] Fix read-write race in serialization, using lock + rename.
Summary:
There was a race in serialization code where a reader could try to deserialize a partially overwritten value.
Here we use a `file.lock` to synchronize the writers, while the readers are not synchronized.
The atomic `rename` is used to move `file.lock` to `file` and make it available to readers.
Since `rename` is atomic, readers should see possibly stale, but not corrupted, data.

Reviewed By: sblackshear

Differential Revision: D4689664

fbshipit-source-id: dc5b546