[infer][backend] remove the perf stats from the summaries
Summary:
It is definitely useful to collect information about how long the analysis of every procedure takes. It allows to detect and focus on outliers when trying to improve performance.  However, this kind of information could be collected using a standard logging mechanism and does not need to be stored within the analysis artifacts.

I intend to add some form of similar logging in the context of #16348004 once we can get every analysis procedure analyzed through the `Ondemand` module. In this case, it would be easy to have a single place to log how does the analysis of a procedure take.

Reviewed By: jberdine

Differential Revision: D4636755

fbshipit-source-id: 01f3bca