[eradicate] Add support for new annotation @PropagatesNullable
Summary:
One limitation of Eradicate is that certain nullability patterns are not expressible using simply the `Nullable` annotation.
One such pattern is using the knowledge that a function returns null when passed null, but returns an object otherwise.
The annotation `PropagatesNullable` is a variant of `Nullable` applied to parameters when their value propagates to the return value.
A method annotated
```
  B m(PropagatesNullable A x) { return x == null ? x : B(x); }
```
indicates that `m` returns null if `x` is null, or an object of class `B` if the argument is not null.
Examples with multiple parameters are in the test cases.

This diff builds some infrastructure for annotation transformers: the example above represents the identity function on nullability annotations.

Reviewed By: jvillard

Differential Revision: D4705938

fbshipit-source-id: 9f6194e