[tests] add a `replace` target to all tests
Summary:
Most of our tests work by comparing the output of some command to a baseline
expected output. It's often needed to update that baseline.

Previously, that was crudely done by attempting to move every foo.exp.test file
to foo.exp. This does not work terribly well, in particular because
foo.exp.test might be stale.

Instead, add a `replace` target to every test that knows how to update the
baseline. This allows custom behaviours too, eg in the differential tests.

Most of the tests include base.make or differential.make, so add a replace target
there. A few tests are completely custom, add a replace target to them too.

Reviewed By: jeremydubreil

Differential Revision: D4754279

fbshipit-source-id: ec34273