[thread-safety] report on write outside sync, read inside sync races
Summary:
This was the one type of races we were not yet reporting (besides ones that use the wrong synchronization :)).

Wrote new utility function to aggregate all accesses by the memory they access.
This makes it easy to say which accesses we should report and what their conflicts are.
Eventually, we can simplify the reporting of other kinds of unsafe accesses using this structure.

Reviewed By: peterogithub

Differential Revision: D4770542

fbshipit-source-id: 96d948e