[make] fix Makefile bug in building models
Summary:
Nuking the specs then building the models was not a great idea: the models do
not look at specs but only at some dummy marker files, eg
infer/lib/specs/c_models, so they don't necessarily realize that they need to
be rebuilt when the specs have been nuked!

One easy workaround would be to also delete the marker files, but then we would
*always* rebuild the models when building infer. Not good.

The solution here is to nuke the specs and marker files only when the clang
dependencies change, then rebuild all the models.

Reviewed By: jberdine, jeremydubreil

Differential Revision: D4781424

fbshipit-source-id: 2d2606e