[toplevel] do not read inferconfig location from the command line
Summary:
Try to read .inferconfig in the current directory, then in .., then in ../..,
etc. This can be overriden with the `INFERCONFIG` environment variable.

This removes the need for two-phase parsing, so clean up that code too.

Paths in .inferconfig are interpreted relative to where .inferconfig is located.
This does not apply to other path-sensitive things like regexpes... this is not
a show stopper because regexpes can account for the fact that infer may be
called from different project roots.

Make sure we fail when .inferconfig exists but cannot be read.

Reviewed By: jberdine

Differential Revision: D4843142

fbshipit-source-id: 340a63f