[quandary] add support for C++ parameter passing modes that differ from Java
Summary:
There are two pointer-related operations you can do in C++ but not Java that we need to support in taint analysis:
(1) `*formal_ptr = ...` when `formal_ptr` is a formal that's a pointer type. Java doesn't have raw pointers, so we didn't need to handle this case.
(2) Passing by reference, which Java also doesn't have (everything is pass-by-value).

Reviewed By: mbouaziz

Differential Revision: D5041246

fbshipit-source-id: 4e8f962