[hil] fix crash when translating C code that indexes string literals like arrays or does pointer arithmetic
Summary:
HIL had only been tested in Java, and it made some assumptions about what array expressions look like (the LHS is always resolvable to an access path) and assignments (the LHS is always an access path) that aren't true in C.
Fixed the code so we won't crash in this case.
Thanks to jeremydubreil for catching this.

Reviewed By: jeremydubreil

Differential Revision: D5047649

fbshipit-source-id: e8484f4