[log] log to a single file with different categories and debug levels
Summary:
Change the API of `Logging` wrt to writing to files and to the console (see
changes in logging.mli).

Write only to one log file: infer-out/log. Prefix each line with the kind of
warning and the PID of the process emitting it. Writing with `O_APPEND` is
atomic so the file should not get garbled by concurrent writes. To get the
output of a single process, find out which one interests you by looking at
infer-out/log, then `grep ^[<PID>] infer-out/log`.

Introduce 3 log levels for debug output and command-line options to set them
for various categories individually.

Change tons of `"\n"` to `"@\n"` so the `Format` module is aware of newlines
without us having to look through every character of every logged string for
`\n` characters.

Reviewed By: mbouaziz

Differential Revision: D5165317

fbshipit-source-id: 93c922f