[buck] do not pass unused --buck to python
Summary:
It is not used, so one fewer thing to track inside Python.

Note that Python does pass `--buck` to OCaml infer to make it aware that it is
running from within Buck, and the OCaml code uses that flag to trim infer-out
to only what's strictly necessary (and especially to remove non-deterministic
output that would mess with Buck's cache).

Reviewed By: jberdine

Differential Revision: D5200043

fbshipit-source-id: 1c84442