[merge] only try to read/create multilink files when --merge is specified
Summary:
Thanks to the logging introduced in D5293334 (or because of, depending on your liking of spam), I noticed that there lots of errors being logged of the form

  Couldn't read multilink file '/home/jul/infer/infer/tests/codetoanalyze/cpp/errors/infer-out/attributes/d0/multilink.txt': /home/jul/infer/infer/tests/codetoanalyze/cpp/errors/infer-out/attributes/d0/multilink.txt: No such file or directory

I don't think it makes sense to care about multilink files except when `--merge` is specified.

Also introduced a .mli and self-documented a function.

Reviewed By: akotulski

Differential Revision: D5310129

fbshipit-source-id: c3a6276