[inferbo] Pruning return value of function call
Summary:
Problem: The analyzer did not know that the value of `v.size()` is an alias of `v.infer_size`, so `v.infer_size` is not pruned by the if condition. As a result it raises a false alarm.

  void safe_access(std::vector<int> v) {
    if (v.size() >= 10) {
      v[9] = 1; // error: BUFFER_OVERRUN Offset: [9, 9] Size: [5, 5]
    }
  }
  void call_safe_access_Good() {
    std::vector<int> v(5, 0);
    safe_access(v);
  }

Solution: Adding alias for return value to the abstract domain.

Now Inferbo can prune `v.infer_size` because it knows that the value of `v.size()` is an alias of `v.infer_size`. There is already an alias domain in Inferbo, so we added a specific room for the retrun value.

Reviewed By: jvillard, mbouaziz

Differential Revision: D5396988

fbshipit-source-id: 4a4702c