[infer][java] Infer should not suggest @Nullable on captured parameters
Summary:
This just makes the warnings silent for now. We may improve the analysis to check if the null check on the captured fields are consistent with the annotation on the corresponding parameters.

Eradicate also has the same issue. I added a test to outline this. The biabduction analysis will also probably fail on the same of annotation lookup. We may want implement the proper fix at the level of `Annotation.field_has_annot`.

Reviewed By: sblackshear

Differential Revision: D5419243

fbshipit-source-id: 6460de8