[thread-safety] make a distinguished access kind for container writes
Summary:
The way we represented container writes before was pretty hacky: just use a dummy field for the name of the method that performs the container write.
This diff introduces a new access kind for container writes that is much more structured.
This will make it easier to soundly handle aliasing between containers and support container reads in the near future.

Reviewed By: da319

Differential Revision: D5465747

fbshipit-source-id: e021ec2