Remove unsupported `index-store-path` argument from clang commands
Summary:
The `-index-store-path` argument does not exist in Clang 5.0 and it gets discarded without any error message. The problem is that its argument, a folder, is not discarded, and Clang considers it as a source file. This leads to the following errors:
- `cannot specify -o when generating multiple output files` (this happens if a `-o` argument is passed)
- `error reading '<PATH>'` (this can be observed when running the "normalized" version of the clang command, generated via the -### flag)

This weird case can be observed when `-index-store-path` is passed in a sequence like the following: `-x c -index-store-path <PATH> -c`.

With this change, we remove the `index-store-path` option, and its argument, from the original clang command.

Reviewed By: jvillard

Differential Revision: D5601808

fbshipit-source-id: 4200308