[filtering] improve issue type filtering CLI
Summary:
Instead of a whitelist and blacklist and default issue types and default
blacklist and filtering, consider a simpler semantics where
1. checkers can be individually turned on or off on the command line
2. most checkers are on by default
3. `--no-filtering` turns all issue types on, but they can then be turned off again by further arguments

This provides a more flexible CLI and is similar to other options in the infer
CLI, where "global" behaviour is generally avoided.

Dynamically created checkers (eg, AL linters) cause some complications in the
implementation but I think the semantics is still clear.

Also change the name of the option to mention "issue types" instead of
"checks", since the latter can be confused with "checkers".

Reviewed By: jberdine

Differential Revision: D5583238

fbshipit-source-id: 21de476