exinferbo] check end of procedure for unreachability
Summary:
Calling exit at the end of a proc does not create unreachable code, prior to this commit inferbo reports that it does.

We extend collect_instrs to detect when we're at the end of a procedure in C and not report on unreachable code if we call a procedure there.

Reviewed By: mbouaziz

Differential Revision: D5623637

fbshipit-source-id: 0d5f326